pyproject.toml
[project]
name = "sul-ai-folio-reporting"
version = "0.1.0"
description = "AI-assisted FOLIO Reporting for library staff"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "folioclient>=1.0.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
